refactor amber hook and fix circ deps
