+++ utils/index.ts
@@ -12,3 +12,4 @@ export * from "./general";
 export * from "./types";
 export * from "./keychainStorage";
 export * from "./signing";
+export * from "./amber";
